move new tabs to replace old tabs
