+++ src/module/vue/components/tabs/tab-helpers.ts
