+++ src/module/vue/components/tabs/tab-list.vue
