+++ src/module/vue/components/tabs/tab-panel.vue
