+++ src/module/vue/components/tabs/tab-panels.vue
